fixup playlist index route
